# String#chomp can be faster than String#sub