--- a/code/string/sub-vs-chomp.rb
+++ b/code/string/sub-vs-chomp.rb
@@ -1,0 +1,17 @@
+require 'benchmark/ips'
+
+SLUG = 'YourSubclassType'
+
+def slow
+  SLUG.sub(/Type$/, '')
+end
+
+def fast
+  SLUG.chomp('Type')
+end
+
+Benchmark.ips do |x|
+  x.report('String#sub/regexp/')   { slow }
+  x.report("String#chomp'string'") { fast }
+  x.compare!
+end