@Arcovion 2015-07-27
How about:
```ruby
SLUG.slice!(/Type$/)
SLUG
```
Probably a fairer comparison.