@bquorning 2015-07-27
Yeah, since `#slice` returns the deleted portion of the input, it only makes sense to compare the “bang” versions. As you say, a separate PR comparing `#slice!`, `#sub!` and `#chomp!`?

I made this PR comparing `.sub` with `.chomp` because I keep seeing the inefficient `.sub` calls in the codebase I’m working on.