@bquorning 2015-11-04
> Sorry took so long to merge this. 

And here I thought this repository was about speed improvements :wink: 

No problem :beers: